Publish change notes along with manual
This commit adds an additional manual draft export step at the point of
publishing a manual. This was done in order to address the problem of
a change note only being sent to the publishing API following
a second, subsequent creation of a manual draft and its publication. In
other words, the change notes always lagged "one step behind" the
publication of the content changes that they described.

By adding a draft export step to the publication of a manual we ensure
that the change notes that were added to the current draft are then sent
through to the publishing API.

Also:
  - Modified some publishing specs to take account of this additional
  draft export step.