Add migration to republish a manual with an older published date
We got a zendesk ticket complaining about the updated at date on a manual
that had been recently added to manuals publisher.  The manual was
originally published in March 2012, and the Dec 2016 version had no
changes other than it appearing on gov.uk, but the updated at date listed
Dec 2016 making some users get in touch with the authors asking what had
changed.

This PR adds a migration that does a one-off republishing of the manual
to:

1. update the first published at dates, and public timestamps to the
   March 2012 date of original publication
2. change the default change notes on the sections to "Imported to GOV.UK"
   instead of "New section added"

In a separate PR we'll add the functionality that would let a user of the
system manage this themselves.