Fixes flappy test for manual_publication_log_filter_spec
Turns out the previous fixes in #783 weren't enough - although we'd
identified that the ordering was the problem, it wasn't the dates that
were the problem.  The test asks for the publication logs to be ordered
by id, but this does nothing.  Although id is a method on a mongoid object
that does indeed return the id of the object, the db field is actually
called `_id`. Asking to sort by `id` does nothing, which means we get
things back in an indeterminate order.

Asking to sort by `_id` gets the data in the expected order and the test
consistently passes (or so it seems anyway).

fixes #789