Add logic to SpecialistDocument#has_ever_been_published?
It's not enough to ask if any editions are published.  When we publish a
manual we change the state first, then we write out publication logs,
then we export a draft, then we tell the publishing-api to publish the
drafts, and then we mark the manual as exported.  When we write out the
final draft it is important to know if the manual or any of its documents
have ever been published before - if we only look at the state of the
documents then for a document with only one edition we would think that
it has been published, when in fact it is in the process of being
published, but has not actually been published.

We change to checking if the document needs exporting for the single
edition case and this means our draft writes out the correct update type.