Provide change note only UI for new sections
It seems sensible to allow useres to edit the change note for a first
edition of a section that is added to a manual.  When we get round to
letting them edit the original publication date, this will allow them
to change the default note from "New section added." to something more
appropriate ("Imported to GOV.UK" perhaps, like we did in
8242435214905c00c41e258a01245d8ce39451e2).

This requires that we tidy up a bunch of the cucumber features to make
it clearer what we expect.