Provide UI for editing originally published at on manuals
For now we present this on the first editions only.  The UI is based on
the similar functionality on detailed guides in whitehall, where the user
can choose to say that the manual is new or was previously published and
this opens up a date select to pick the time.  In order to parse the
date and time data into a DateTime object to store on the models we have
to replicate some of the activerecord/mongoid multiparam attribute handling in
the controller because of the layers of indirection between the controller
and the actual persisted object.