Add toggle to Manuals to control public timestamp
If the manual has an orignally_published_at field it's clear what we
should do when publishing the first edition - we set the
first_published_at and public_updated_at to this value.  It's less
clear what we should do on future editions; for future minor editions
we could do the same, but for future major editions or future minor
editions after major edition other than the first one.  We could
probably try to calculate what to do, but MP makes it hard to work
with previous editions of a manual at the points where we'd need to
make these descisions.

To avoid all this we leave it up to the user - we'll give them a UI
to explicitly say if the originally published date should be used as
the public updated at on GOV.UK, or if it should only set it as the
first published at and let publishing api set the public timestamp
itself.

This commit adds the plumbing for the toggle on the manual editions
and forces the value to "true" when creating a new manual.