Add UI to let the user edit original publication date later
If the original publication date, or the
use_originally_published_at_for_public_timestamp toggle changes we need
to write out new drafts of the whole manual including its sections. It's
hard to detect what has changed during a normal update event because of
the abstraction in the manual models where we don't have access to previous
editions.  To get around this we make editing the date a separate UI and
action so that we can always do this (even if the update turns out to be
a no-op).

Note that this is unlike other updates in that we have to force a new
edition of all the documents so that we get a new draft written, even
if they otherwise wouldn't write this draft.  We don't need to make any
other changes, so we issue an empty update.