Allow for clearing of the originally published date
If the "previously_published" toggle is set to "0" then we nil out the
"orignally_published_at" date param sent in the params hash in the
controller.  This probably isn't the best place to do it, but it means
all the params massaging (multiparam -> single value, etc...) happens
in the controller layer, and the service layers can be simpler and
not have to worry about this.