Add migration to add original publication date to a manual
This is the national planning policy framework manual that we had to set
the first_published_at and public_updated_at timestamps of in #788.  We
set the originally_published_at and use_originally_published_at_for_public_timestamp
values that we would have used in #788 had they been present.  One
minor issue is that we have to set the toggle to true on the latest
edition in order to make sure that a republish would do the right thing,
this does mean however that any future edition would inherit that value and
persist the older date.