Send update_type to exporters when the action type is republish
The action param for the callbacks in ManualObserversRegistry is set to
`:republish` when notified from the RepublishManualService.  When this
is the case we want to send `update_type: republish` to the exporters
(when writing the draft and when publishing the content) so that we
tell publishing-api that we're doing a republish event, not a standard
content change and publish.