Add VersionedManualRepository
This repo doesn't have the same interface as the existing repositories,
but as it is intended to be used in a different way I don't think that's
necessarily a bad thing.

This repo, given a manual id will return a hash containing a published
key and a draft key.  Either (or both) of these could be nil, although
the repo will throw and exception if the manual id doesn't refer to an
actual manual.

The simple case is for a manual that has yet to be published - the
published key will be empty, and the draft key will contain the Manual
instance that represents the supplied id, with the most recent version
of all its specialist documents attached.  If the manual has been
published, but has no draft its equally simple - the draft key is empty
and the published key contains the Manual instance.  In the former case
everything should be in "draft" state, in the latter everything should
be in "published" state.

It gets complex when the manual has been published and has a new draft
ready.  In this case both keys are filled.  The draft key has the most
recent version of the manual and its documents are the most recent
version of each.  The state of the manual will be "draft" but the state
of the documents could be "draft" or "published" (depending on if they
have been edited since the manual was published or not).  The published
key has a Manual instance that represents the previous version of the
manual and its documents are the most recent published version of each.
The state of the manual and all its documents should be "published" in
this case.

If the current or previous version of the manual is "withdrawn" it is
not included in the returned hash as either a "draft" or "published"
version.

Note that because the data model is lossly, we can't correctly obtain
arbitrary versions of a manual and the correct versions of its documents.
The data model maintains a link between a manual edition and the document
editions, but it does not maintain a version link.  All we can say is
that version N of manual X has document A in it, but not which version of
document A.  This means we can only say with confidence what versions are
in the first edition of a manual, the current edition of a manual, and if
the current edition is draft the previous edition of the manual.

Also worth noting is that the Manual and specialistDocument instances
returned by this repo are "read only" in that they don't have functioning
"document_builder" or "slug_generator" factories.  For our purposes this
is ok (and easier than wading through the abstraction to instantiate the
correct instances).