Use the versioned manual repository in the manual relocator
The manual relocator already had to work with published and draft
versions of a manual in order to do the right thing when moving editions
around. It previously hand-rolled its efforts and avoided using the
Manual and SpecialistDocument models.  By using the new repo we can use
these models and delete our own "Simple" implementations.