Remove usages of SpecialistDocumentEdition#extra_fields
It looks as if the means to edit this field via the UI was removed in
the giant "Remove specialist document models" commit [1]. As far as
I can tell this field was only ever needed for documents which are
now published via the Specialist Publisher app. I believe we could
also remove the field itself. However, this is complicated by:

a) There are some instances of SpecialistDocumentEdition in the local
data store which have the `extra_fields` attribute set. At first glance
it looks as if these might all be orphans and in any case relate to
content items which are now published via the Specialist Publisher app.

b) There are some existing migrations which rely on this field.

I have not addressed those issues in this commit.

[1]: 3162f0e72fd53bfd1cf87f5661bb198ff785940e