Make republishing work for draft as well as published manuals
We use the new VersionsManualRepository to grab both the draft and
published versions of the supplied manual_id.  We've also changed the
service to take draft vs. published listeners.  If the supplied manual
has a published version we notify the published listeners and if it has a
draft version we notify the draft listeners.  In both cases we send the
"republish" action to make sure the listeners do the right thing when
talking to the publishing-api.