Upgrade to newer cucumber to make it work on ruby 2.2
We ran rails g cucumber:install and took the opportunity to remove all
the custom entries from env.rb so it's easier to upgrade in the future
without worrying about differences in that file.  The custom code goes
into separate files in features/support as cucumber loads all files in
here by default so we don't have to do it explictly.

One gotcha is that we have to make sure that rspec/mocks is included
before what used to live in api_helpers to stub rummager.  As this isn't
shared between rspec and cucumber we remove api_helpers (and simplify
the spec_helper in the progress) and do what it used to do in world_setup
which we can be sure has been required after the mocks setup.