Upgrade factory_girl
We swap to using factory_girl_rails which does the automatic factory
loading for us so we can delete some of our hand-roled stuff.  We also
drop the inclusion of the FactoryGirl syntax methods from our rspec
tests. This means we have to use FactoryGirl.create instead of invoking
create directly, but because we hadn't done that for our cucumber steps
and a lot of our helpers are shared, there were only a few places we
were using that syntax.