Add fallback for asset_manager_bearer_token when not present in env
Without a non-nil value for this the asset manager on dev will reject all
requests.  Due to a bug in the implementation of asset saving this means
the system silently fails in dev mode when trying to upload an asset and
doesn't break until you save the section to the publishing-api.