Drop un-needed explicit Gemfile version requirements
We prefer explicit Gemfile version requirements for libraries only when
there's a reason not to upgrade further (e.g. other dependencies might
limit the upgrade).  This goes doubly for tools like webmock, foreman,
database_cleaner, etc... where it's not usually a problem to be upgraded
and these usually only affect development and test.

We also move foreman to be a development and test gem following the
pattern in specialist-publisher. Although the app does have a Procfile for
use on deployed environments, we don't use foreman to deal with running
the app.