Bump gds gems as far as we can
Going further on many of these would require rails 4 or mongoid 5 so
we're just going as far as we can for now.  For govuk_frontend_toolkit
we go as far as the same version that manuals-frontend has; it's good
to keep these in sync as we want the previews of markdown that
manuals-publisher does to be as close as possible to the "live" display
we'd get from manuals-frontend.