Remove SpecialistDocumentEdition#extra_fields field
This field was only ever used for non-manual specialist documents,
but these are now stored in the Specialist Publisher app. So the
field is no longer needed and can be removed.

There hasn't been a way to set this field via the user interface
since the giant "Remove specialist document models" commit [1].

I've run the following queries against the production data dumped on
2017-02-23 before and after the 20170224113808 migration to confirm
that there will be no SpecialistDocumentEditions with this field set
to anything other than an empty Hash after the migration is run:

    > SpecialistDocumentEdition.count
    => 26704
    > SpecialistDocumentEdition.where(extra_fields: {}).count
    => 3711
    > SpecialistDocumentEdition.where(:extra_fields.ne => {}).count
    => 22993

    $ rake db:migrate
    ==  RemoveSpecialistDocumentEditionsWithNonManualDocumentType: migrating ======
    ==  RemoveSpecialistDocumentEditionsWithNonManualDocumentType: migrated (0.7991s)

    > SpecialistDocumentEdition.count
    => 3686
    > SpecialistDocumentEdition.where(extra_fields: {}).count
    => 3686
    > SpecialistDocumentEdition.where(:extra_fields.ne => {}).count
    => 0

The field is still referenced in some old migrations, but don't think
there's any need to modify them.

[1]: 3162f0e72fd53bfd1cf87f5661bb198ff785940e