Require test-unit gem for all environments
We only really need it for running tests, and only because we're using
rails 3.2 (which relies on test-unit) and ruby 2.2 (which no longer
bundles a test-unit shim).  However, if it's in the development or test
groups it isn't installed on our deployed environments and this somehow
stops the rails console from booting (but not the server).

Moving it outside these groups makes the console work again.