Rename SpecialistDocumentEdition -> SectionEdition
Although we change the name of the collection that attachments are
embedded in (with `embedded_in :section_edition` in
`app/models/attachment.rb`) we don't need to migrate any data
here. The combination of `embedded_in` on attachments and
`embeds_many` on section editions means that a record in the
underlying mongodb section edition collection looks like:

    { _id: 'abcd1234',
      ... some other fields ...,
      attachments: [
        {_id: 'efgh5678',
         filename: 'foo.txt'
         ... some other fields ...
         }
      ]
    }

That is, the relationship is defined by embedding Attachment records
in an `attachments` array under whatever is defined as the parent
collection. This continues to work when we rename the parent
collection.