Remove `bin/fix_inconsistent_edition_states`
We noticed this script when renaming
`specialist_document_editions`. As far as we can tell it was
introduced in c3fa17e56 and the behaviour hasn't changed since. The
embedded tests at the bottom of the file do not run on current master,
and are not part of the automated build.