Use fork of mongoid_rails_migrations to fix build
* This is an attempt to fix the problem described in #816. It uses a
forked version of mongoid_rails_migrations. The branch in that fork
was branched off the v1.0.0 tag, which is the version of the gem that
was specified in our `Gemfile`. This is so that we keep the changes to
our app to a minimum.

* This fix is a bit of a bodge - the proper fix would be along the
lines of adacosta/mongoid_rails_migrations#3, but given that the failing
build is a significant blocker, it feels as if this is worth a try.
If it works OK, we can work on getting the proper fix in place.

* I installed v1.14.5 of Bundler which is the version now specified in
govuk-puppet [1] for Ruby v2.2.4. Hence the change to `BUNDLED WITH`
in the `Gemfile`.

* I've run a migration locally to ensure that the gem still works OK
even with this patch in place.

[1]: https://github.com/alphagov/govuk-puppet/blob/e2d61705aaf6da180a781b6f46f83235fd890df0/modules/govuk_rbenv/manifests/all.pp#L46