Extract magic strings into PUBLISHING_API_XXX constants
I want to remove `SpecialistDocumentEdition#document_type` which is now
always set to 'manual'. In this commit I'm trying to make it clearer
that the 'manual' strings used in `ManualPublishingAPIExporter` are not
related to `SpecialistDocumentEdition#document_type`; in this case
they're the name of the schema and the document type sent to the
Publishing API and are not used *internally* within this app.