Remove unused & broken ManualsPublisher module
Since my last commit, this module is no longer used anywhere and in
any case it is broken in a couple of ways:

* The `AbstractDocumentServiceRegistry` class was removed in this
commit [1].
* The `ManualsPublisher#observer_registry` method was removed in this
commit [2].

I've also had to remove a `require` statement at the top of
`config/routes.rb` to get the tests to pass. This require statement
became redundant in this commit [3], but was not removed at that point.

[1]: 43e2e7534da179f92bed2ce7e73bc3f3c4fdfb06
[2]: 9bbf2d34128fc53c5ead049240cc64d732b652f9
[3]: 9e6b913343a7735b8ff0bc9a5999d9582bc85165