Use generic organisations in manual_record_spec
I think the CMA organisation might have been significant in the past.
It's no longer significant in this app as CMA Cases are now published by
specialist-publisher[1]. Using generic organisations in this test should
make it clearer that we only care that the orgs are different, and not
that they have specific slugs.

[1]: https://github.com/alphagov/specialist-publisher