Fail fast if uploading an attachment responds with 404
I was seeing an exception in development when trying to save a Section
with an attachment. I eventually realised that it was because I didn't
have the Asset Manager running and therefore the attachment hadn't
actually been uploaded. There was no indication of this problem in the
user interface.  It looks as though the call to `errors.add` was
intended to do something sensible in the case of an error but it doesn't
appear to have any effect. The result is that errors are essentially
swallowed and the user is led to believe that the attachment has been
uploaded successfuly.

We can't assume that a 404 error means that the Asset Manager is
definitely unavailable as the API will return 404 in response to a call
to `#update_asset` with a file_id that doesn't exist. Thanks to
@floehopper for prompting me to investigate this behvaiour.

In order to add a test for the new behaviour I've made the
`#upload_file` method public.