Move rummager to Services module
Using a `Services` module to keep all the GdsApi clients together is a
pattern used in a number of other GOV.UK apps.

I'm manually requiring "services" where appropriate as the lib directory
isn't in the array of `autoload_paths`.

Note that we're already requiring "services" in section_reslugger.rb so
I haven't had to require it again.