Combine ManualDocumentRenderer and SectionRenderer
Fixes #810

In 4c4fe432e we renamed `SpecialistDocumentRenderer` to
`SectionRenderer`. We have since noticed that Sections are
sometimes referred to as `ManualDocument` in the code. This commit is
part of a larger renaming refactor to replace references to 'manual
document' with 'section'.

In doing so I noticed that we have two renderers for section, the
former `SpecialistDocumentRenderer` (before this commit `SectionRenderer`)
and `ManualDocumentRenderer`. As far
as I can tell the only difference is that the former was used for
previewing changes to sections and did not include footnote rendering
as part of the pipeline.

This seems like an oversight, as the rendering of sections for citizen
consumption does use the footnote rendering - it makes sense for the
preview to also do this. I've taken the opportunity therefore to use a
single `SectionRenderer` throughout.