Remove export_recipient param from SectionPublishingAPIExporter
The `export_recipient` is always the `PublishingApiV2#put_content`
method. Setting it in the class rather than passing it to the
constructor makes this easier to see.

I've had to update the "sets it to major if the document has never been
published" example to stub a method on `document` instead of changing a
value in the `update_type_attributes` hash. This is because the
`document` is now instantiated in a `before` block (in the call to
`allow(subject)` before we set the value of `ever_been_published` in the
hash.