Extract ManualWithdrawer class from withdraw_manual script
I've chosen to pass in a logger rather than using stdout/stderr to avoid
output when running the tests. It also makes the script more consistent
with the `republish_manuals` script.

I've also chosen to raise an un-handled exception rather than calling
`Kernel#exit` directly, because it felt even less sensible to do this
from within the method of a class. If the exception is not handled this
will result in the same non-zero error code.