Remove ManualRelocator
The last reference to this class was removed in f8698ce22. It is
mentioned in #851 as a recent example of useful script that has tests
so we may want to use it in the future to base a similar migration
on. But for now it cannot be used externally via a script or rake task
so I think it is safe to remove it.