Remove export_recipient param from ManualPublishingAPIExporter
The `export_recipient` is always the `PublishingApiV2#put_content`
method. Setting it in the class rather than passing it to the
constructor makes this easier to see.

I've had to update the setup in the shared examples for "publishing a
manual that has never been published" to stub a method on `manual`
instead of changing a value in the `manual_attributes` hash.  This is
because the `manual` is now instantiated in a `before` block (in the
call to `allow(subject)` before we set the value of
`ever_been_published` in the hash.