Rename routes from `manual_document` -> `section`
This commit changes the routes so that `sections` rather than
`manual_documents` is the name of the nested route under `manuals`.

Because `attachments` is also nested under `sections` all of the
methods that generate URLs from the routes config for both attachments
and sections need to be changed at the same time in this commit.

Note that the code uses a `SectionViewAdaptor` which acts like an
active record model. When it is passed to certain helpers such as
`form_for` it is used to generate both URLs and HTML. Because of the
HTML generation I have had to change a few locations in the code where
the structure of the generated HTML is assumed - notably in
`app/views/sections/_form.html.erb` where the presence of a DOM node
with the id `#section_body` is hard-coded.

In a subsequent commit I will add a test to ensure that the generated
HTML includes the DOM elements that the JS expects.