Inline listeners in RemoveSectionService
This service is only called from one location and with a fixed set of
listeners, so these can be inlined. The specs relied on the dependency
injection for testing, so I've had to replace that with a stub on the
methods that are called on the services collaborators. This makes the
specs slightly more verbose but at least makes it clear which API
methods get called as a result of using this service.