Use GdsApi::Response#[] instead of #format
It looks as though we should be treating a `GdsApi::Response` as a hash
instead of an object with methods.

This is the same change that has been made in PR #863.

I _think_ that the `format` attribute has also been deprecated/removed
but this change is good enough to get past this particular problem in
the script. I'll work out whether it's better to use `document_type` or
`schema_name` later.

I tested this change in the Rails console:

    require 'services'

    item = Services.content_store.content_item('/guidance/manual-123')

    item.format
    NoMethodError: private method `format' called for #<GdsApi::Response:0x007f9e136992c8>

    item['format']
    => "manual"