Rescue ItemNotFound from Content Store
I _think_ the `validate_new_section_in_content_store` method is there to
ensure there's no content with our target slug in the content store. An
ItemNotFound indicates that it doesn't exist so we should be good to
continue.

I think ideally there'd be an API method we can use to check whether
content exists but that doesn't seem to be the case at the moment.

Note that I have to instruct Rubocop to ignore the fact that we're
rescuing an exception but not doing anything with it.