Simplify Organisations API initializer
There is already a mechanism available for overriding the URI for
specific services, so there is no need to have some special code to do
this. Service URI's are automatically looked up via `Plek` and they can
be overridden using environment variables of the form:
`PLEK_SERVICE_<service-name>_URI` (see [1]).

In this case, we can set the `PLEK_SERVICE_WHITEHALL_ADMIN_URI`
environment variable to "https://www.gov.uk" if we want to use the
production website to supply the data for the Worldwide API in
development.

The reference to the "preview" environment is out-of-date; I think this
is now the "integration" environment and that will have an instance of
"whitehall-admin" running, so I think we can ignore it.

c.f. https://github.com/alphagov/smart-answers/commit/9c01ec59eb2a715cfdf641a7db85e1b84d03677c

[1]: https://github.com/alphagov/plek/blob/b1f917cd6498aea8867728dab0132d0373ca702b/README.md#for-base-urls