Fix PublishingAPIRedirecter for v2 of Publishing API
I determined the attributes to export by reading the examples[1],
validating against the schema in the spec and by trying this code
against a local version of the Publishing API in the Rails console.

The only attribute I haven't set is `public_updated_at` as it doesn't
appear to be required.

[1]:
https://github.com/alphagov/govuk-content-schemas/blob/master/formats/redirect/publisher_v2/examples/redirect.json