Inline ManualObserversRegistry#publication
This method on the ManualObserversRegistry is only called from the
PublishManualService. I think it makes it more obvious what the
service does if these actions are part of the service itself.

This commit removes a comment in the `publication` method on
ManualObserversRegistry which stated:

    The order here is important. For example content exporting should
    happen before publishing to search.

    A draft export step follows publication logging as this ensures
    that change notes that relate to the current draft are pushed
    straight to the publishing API rather than on the subsequent
    draft-publish cycle.

I've decided to replace this comment with some new assertions in the
spec for PublishManualService which ensures these collaborators are
called, and called in this specific order. I haven't spent any time
establishing whether it really is necessary to call these
collaborators in this specific order but at least adding them to the
spec makes it less likely that we'd accidentally change this than
relying on the comment alone.