Inline ManualObserversRegistry#republication
The `republication` method on `ManualObserversRegistry` is only called
from the RepublishManualService so inlining it there makes it more
obvious what that service does.

I haven't preserved the comment on this method in
ManualObserversRegistry as I think this change makes it more obvious
what method arguments each of the classes are called with.