Inline ManualObserversRegistry#update
This method was called from two places so although it adds a little
bit of duplication, I've inlined the method call into the two services
that call it in order to make it more obvious what those services
do. In the case of the RepublishManualService this also makes the
specs more explicit in their testing of which api calls get made as a
result of republishing a manual.

As this also removes a method that is aliased, I've removed the alias
and renamed the method.