Extract MarkedSectionDeleter
I've moved the methods from the script to the class and the rest of the
code into an `execute` method. This follows the pattern that @floehopper
has used when extracting other classes from bin scripts (e.g.
825771c9aa2e2d3228761d473412fe6c3e7354cf).

I'm passing a logger to the initializer so that we can avoid polluting
test output.