Add missing require statements
I needed to add these in order to be able to navigate to /manuals/new in
development.

They've been needed since 30a72a16dadde70c86a669ae23f715f20554b7fd.
Prior to that commit, constructing a `RepositoryRegistry` (which happens
indirectly in `ManualsController#index`) would result in a
`DocumentFactoryRegistry` being created which in turn would result in
the necessary builders and validators being required.

As of this commit I appear to be able to do everything through the user
interface as expected.

A longer term fix might be to autoload/require the builders and
validators as suggested in issue #907.