Add `app/models/builders` to `autoload_paths`
Before this commit, running

    `be rails r 'puts ActiveSupport::Dependencies.autoload_paths'`

Shows that the version of rails we are using does not add
subdirectories under `app` to the `autoload_paths` by default. This
means we have to remember to require files in `app/models/builder` in
order for the app to work correctly in development. This does not
affect production as we have `config.cache_classes = true` set in
`config/environments/production.rb`.

This commit adds `add/models/builders` so we don't have to remember to
require these files.