Improve tests for MarkedSectionDeleter
I've added tests to broadly describe the current behaviour of
MarkedSectionDeleter. There are certainly cases that won't be covered
but I think these should serve as a good starting point should anyone
else need to make changes to the class.

Note that I'm only testing behaviour that was already present. I can't
vouch for the correctness of the behaviour itself.