Update production config in mongoid.yml
These were the options that were removed when I regenerated the
mongoid.yml file in commit titled "Add auto-generated mongoid.yml":

    uri: <%= ENV['MONGODB_URI'] %>
    database: govuk_content_production
    refresh_mode: :sync
    logger: false
    refresh_interval: 120
    use_activesupport_time_zone: true

Note the absence of `database` in the mongoid.yml production config.
This is no longer required as the MONGODB_URI environment variable
already includes the database name. The database name is set to
'govuk_content_production' in backend.yml[1] and used to populate the
MONGODB_URI in manuals_publisher.pp[2]. The MONGODB_URI string is built
in mongo_db_uri.pp[3]

This production config was initially added in commit
c371b95c14bb1ada89dc6fd31afaefdf3a154f42 (16 Sep 2016) but doesn't
contain any explanation about the options.

The `refresh_interval` appears to be passed on to `Moped::Cluster`.

The `logger` option appears to be used by Mongoid.

The `use_activesupport_time_zone` option appears to be used by Mongoid.

The `refresh_mode` option doesn't appear to be available in Moped
although I haven't been able to find any mention of its absence. The
option also appears to have been removed from the Mongo Ruby Driver in
1fc62d949269231e8e7639589a336276390dcc65 although it's not mentioned
specifically in that commit. This blog post[4] explains that some drivers
need to be initialized to refresh the view of the replica set:

> The Ruby driver (as of 1.9.2), for example, does not refresh its view
> of the replica set unless the client is initialized explicitly to do
> so with refresh_mode: :sync. Check your driver documentation.

I wonder whether the implication there is that this isn't required for
more recent drivers. That would explain the absence of the setting in
Moped and later versions of the Mongo Ruby driver.

[1]: https://github.com/alphagov/govuk-puppet/blob/5b422711e5fe20cb4c37f260fc8a7b362f94d055/hieradata/class/backend.yaml#L19
[2]: https://github.com/alphagov/govuk-puppet/blob/master/modules/govuk/manifests/apps/manuals_publisher.pp#L86
[3]: https://github.com/alphagov/govuk-puppet/blob/a6c51d887a6501f02766b7279127b60f02037a7f/modules/govuk/manifests/app/envvar/mongodb_uri.pp
[4]: https://devops.com/mongodb-replication-pro-tips/#somedriversrequiremanualsync