Update call to store_in for Mongoid 3
This addresses the following error message I was seeing when running
`rake`:

    Problem:
      Invalid options passed to SectionEdition.store_in:
      manual_section_editions.
    Summary:
      The :store_in macro takes only a hash of parameters with the keys
      :database, :collection, or :session.
    Resolution:
      Change the options passed to store_in to match the documented API, and
      ensure all keys in the options hash are symbols.

Note that we're already passing a hash to `store_in` in the User model.
I'm fairly certain this wasn't working as expected prior to the upgrade
to Mongoid 3 as the first param of the `store_in` method was a string
containing the name of the collection. Passing a hash appears to have
resulted in the users being stored in a collection named
“{:collection=>:manuals_publisher_users}” in development. Note that this
wouldn't have affected the production database as those users are stored
in the single sign on app.